interpreted-runtime: mock Node.getId() with a long, not a boxed Int ()
When interpreted-runtime was migrated to Scala 3, the compiler stopped
implicitly widening Int to Long. The test now fails with

An exception or error caused a run to abort: Integer cannot be returned
by getId()

during initialization. This fact is hidden by the fact that the test
runner ignores tests that fail discovery.

+++ community/cypher/interpreted-runtime/src/test/scala/org/neo4j/cypher/internal/runtime/interpreted/pipes/NodeHashJoinPipeTestSupport.scala
@@ -62,7 +62,7 @@ trait NodeHashJoinPipeTestSupport extends InterpretedRuntimeTestSuite {
     }
   }
 
-  protected def newMockedNode(id: Int) = {
+  protected def newMockedNode(id: Long) = {
     val node = mock[Node]
     when(node.getId).thenReturn(id)
     when(node.getElementId).thenReturn(id.toString)
